+++ lib/events/entity_add_event.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_pom/events/entity_change_event.dart';
+import 'package:flutter_pom/flutter_pom.dart';
+
+class EntityAddEvent<T extends Table> extends EntityChangeEvent {
+
+}